VCard: Ignore lines causing invalid line exception
When parsing VCard data that has malformed property data the parser
will throw a VCardInvalidLineException. However, further up the
parser this exception is not handled and because of this the VCard
parser will quit, causing NO VCard data to be set as contacts. This
can be mitigated by ignoring the exceptions caused by malformed
property data. By doing so, the VCards with malformed property data
will still be generated and the parser will continue to parse the
rest of the VCard data that follows.

Details:
This change will ignore the VCardInvalidLineException thrown by
constructPropertyData when it is unable to parse a VCard property. The
property will be ignored, but the VCard parsing will still continue.
This makes the assumption that the END:VCARD line is always a valid
line.

This will also ignore any properties with invalid property names
without throwing an exception, ignoring that property line.  This
makes the assumption that the following line is not a continuation of
the line with the invalid property name.  If so, a
VCardInvalidLineException or another invalid property line will
occur, both of which will be ignored but output a warning log.

This patch fixes b/308970822

Test: atest AndroidVCardTests
Change-Id: If99804cddd2744e8fd815d19aace3266bafbbb91
(cherry picked from commit 3d6914590c0464678fd965fe83ed0fc2e46a66f4)