Update errorprone and fix most of them
Putting errorprone up to par with packages/modules/Bluetooth.

StringCaseLocaleUsage is not enforced as error as 22 warning need to be
fixed.
NonApiType is disabled as external API required an ArrayList directly
UnicodeEscape is left as is, it is not something we fixed in client code
and it was already a warning before

Bug: 311772251
Test: None
Flag: EXEMPT refactor
Change-Id: I7eb8fc31f05a668ee28dbca8b0a2a0f9ff41bd73